DLS-7522 check controllers and add requireData to missing controllers.  Update corresponding spec files
