+++ app/controllers/HowManyCreditsForLostDamagedController.scala
@@ -18,9 +18,7 @@ package controllers
 
 import controllers.actions._
 import forms.HowManyCreditsForLostDamagedFormProvider
-
-import javax.inject.Inject
-import models.{Mode, UserAnswers}
+import models.Mode
 import navigation.Navigator
 import pages.HowManyCreditsForLostDamagedPage
 import play.api.i18n.{I18nSupport, MessagesApi}
@@ -29,6 +27,7 @@ import repositories.SessionRepository
 import uk.gov.hmrc.play.bootstrap.frontend.controller.FrontendBaseController
 import views.html.HowManyCreditsForLostDamagedView
 
+import javax.inject.Inject
 import scala.concurrent.{ExecutionContext, Future}
 
 class HowManyCreditsForLostDamagedController @Inject()(
@@ -45,10 +44,10 @@ class HowManyCreditsForLostDamagedController @Inject()(
 
   val form = formProvider()
 
-  def onPageLoad(mode: Mode): Action[AnyContent] = (identify andThen getData) {
+  def onPageLoad(mode: Mode): Action[AnyContent] = (identify andThen getData andThen requireData) {
     implicit request =>
 
-      val preparedForm = request.userAnswers.flatMap(_.get(HowManyCreditsForLostDamagedPage)) match {
+      val preparedForm = request.userAnswers.get(HowManyCreditsForLostDamagedPage) match {
         case None => form
         case Some(value) => form.fill(value)
       }
@@ -56,16 +55,14 @@ class HowManyCreditsForLostDamagedController @Inject()(
       Ok(view(preparedForm, mode))
   }
 
-  def onSubmit(mode: Mode): Action[AnyContent] = (identify andThen getData).async {
+  def onSubmit(mode: Mode): Action[AnyContent] = (identify andThen getData andThen requireData).async {
     implicit request =>
-
-      val answers = request.userAnswers.getOrElse(UserAnswers(id = request.sdilEnrolment))
       form.bindFromRequest().fold(
         formWithErrors =>
           Future.successful(BadRequest(view(formWithErrors, mode))),
         value =>
           for {
-            updatedAnswers <- Future.fromTry(answers.set(HowManyCreditsForLostDamagedPage, value))
+            updatedAnswers <- Future.fromTry(request.userAnswers.set(HowManyCreditsForLostDamagedPage, value))
             _              <- sessionRepository.set(updatedAnswers)
           } yield Redirect(navigator.nextPage(HowManyCreditsForLostDamagedPage, mode, updatedAnswers))
       )
