+++ app/controllers/KeepAliveController.scala
@@ -16,7 +16,7 @@
 
 package controllers
 
-import controllers.actions.{DataRetrievalAction, IdentifierAction}
+import controllers.actions.{DataRequiredAction, DataRetrievalAction, IdentifierAction}
 import play.api.mvc.{Action, AnyContent, MessagesControllerComponents}
 import repositories.SessionRepository
 import uk.gov.hmrc.play.bootstrap.frontend.controller.FrontendBaseController
@@ -28,16 +28,13 @@ class KeepAliveController @Inject()(
                                      val controllerComponents: MessagesControllerComponents,
                                      identify: IdentifierAction,
                                      getData: DataRetrievalAction,
+                                     requireData: DataRequiredAction,
                                      sessionRepository: SessionRepository
                                    )(implicit ec: ExecutionContext) extends FrontendBaseController {
+def keepAlive: Action[AnyContent] = (identify andThen getData andThen requireData).async {
+  implicit request =>
+    sessionRepository.keepAlive(request.userAnswers.id).map(_ => Ok)
+}
 
-  def keepAlive: Action[AnyContent] = (identify andThen getData).async {
-    implicit request =>
-      request.userAnswers
-        .map {
-          answers =>
-            sessionRepository.keepAlive(answers.id).map(_ => Ok)
-        }
-        .getOrElse(Future.successful(Ok))
-  }
+    Future.successful(Ok)
 }
